// constants/types.ts
export type Artifact = {
  title: string;
  medium: string;
  objectID: string;
  objectURL: string;
  primaryImage: string;
  primaryImageSmall: string;
  objectName: string;
  region: string;
  culture: string;
  artistAlphaSort: string;
  artistPrefix: string;
  artistNationality: string;
  objectDate: string;
};